permissions: add usePermissionsDescriptions - fix ConnectedAccountsPermissions eth_accounts i18n
